get price data from online
Get price data from every card set in the dataset from mtgprice.com

This website was chosen because is has a "fair trade price" which takes into account availability of cards on ebay,amazon and other 3rd party retailers.

i.e: a card being sold many times for a lot of money will have a bigger weight than a card almost never being sold. (as it is not as desirable).

We also take information related to a simple average of the price it is being sold for on various websites.
We are not yet sure which of these values will more accurately portray "performance" of the card.